Corrected string marshaling in .Net Core
On Windows the Ansi conversion is the local code page, not utf-8 as on
Mac/Linux.   So on Windows string with non Ascii characters would be
mangled.

The .Net Core Marshal to UTF8 functions are only defined for the
CoTaskMem allocator so for consistency all marshaling was moved
to that allocator.